@@ -1,19 +1,18 @@
 buildscript {
-    ext.kotlin_version = '1.4.31'
-    ext {
-        compileSdkVersion = 30
-        targetSdkVersion = 30
-        appCompatVersion = "1.0.2"
-    }
-
+    ext.kotlin_version = '1.7.10'
     repositories {
         google()
-        jcenter()
+        mavenCentral()
+    }
+
+    ext {
+        compileSdkVersion = 32
+        targetSdkVersion = 32
+        appCompatVersion = "1.4.2"
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.3'
-        classpath 'com.google.gms:google-services:4.3.5'
+        classpath 'com.android.tools.build:gradle:7.2.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -21,8 +20,9 @@ buildscript {
 allprojects {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
         maven {
+            // [required] background_fetch
             url "${project(':background_fetch').projectDir}/libs"
         }
     }
