@@ -4,19 +4,15 @@ import android.Manifest
 import android.accounts.Account
 import android.accounts.AccountManager
 import android.app.Activity
-import android.app.NotificationManager
 import android.appwidget.AppWidgetManager
 import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
-import android.os.Build
 import android.provider.AlarmClock
-import android.provider.Settings
-import android.widget.Toast
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import androidx.multidex.MultiDex
 import com.transistorsoft.flutter.backgroundfetch.HeadlessTask
-import fr.skyost.timetable.notification.LessonNotificationModeManager
 import fr.skyost.timetable.utils.Utils
 import fr.skyost.timetable.widget.TodayWidgetReceiver
 import io.flutter.app.FlutterApplication
@@ -27,15 +23,12 @@ class UnicaenTimetableApplication : FlutterApplication() {
     companion object {
         const val CHANNEL = "fr.skyost.timetable"
 
-        const val PREFERENCES_FILE = "preferences"
-        const val PREFERENCES_LESSON_NOTIFICATION_MODE = "lesson-notification-mode"
-
         fun handleMethod(call: MethodCall, result: MethodChannel.Result, context: Context) {
             when (call.method) {
                 "account.get" -> {
                     val manager: AccountManager = AccountManager.get(context)
                     val accounts: Array<Account> = manager.getAccountsByType(context.getString(R.string.account_type_authority))
-                    if (!accounts.isNullOrEmpty()) {
+                    if (accounts.isNotEmpty()) {
                         val account: Account = accounts.first()
                         val user: MutableMap<String, Any> = HashMap()
                         user["username"] = account.name
@@ -58,7 +51,7 @@ class UnicaenTimetableApplication : FlutterApplication() {
                 "account.remove" -> {
                     val manager: AccountManager = AccountManager.get(context)
                     val accounts: Array<Account> = manager.getAccountsByType(context.getString(R.string.account_type_authority))
-                    if (!accounts.isNullOrEmpty()) {
+                    if (accounts.isNotEmpty()) {
                         Utils.removeAccount(manager, accounts.first(), result)
                         return
                     }
@@ -70,29 +63,6 @@ class UnicaenTimetableApplication : FlutterApplication() {
                     updateIntent.putExtra(TodayWidgetReceiver.INTENT_REFRESH_WIDGETS, true)
                     context.sendBroadcast(updateIntent)
                     result.success(null)
-                }
-                "activity.lesson_notification_mode_changed" -> {
-                    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
-                        return
-                    }
-
-                    val mode: Int = call.argument<Int>("value")!!
-                    context.getSharedPreferences(PREFERENCES_FILE, Context.MODE_PRIVATE).edit().putInt(PREFERENCES_LESSON_NOTIFICATION_MODE, mode).commit()
-                    if (mode != LessonNotificationModeManager.VALUE_DISABLED) {
-                        val manager: NotificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && !manager.isNotificationPolicyAccessGranted && context is Activity) {
-                            Toast.makeText(context, R.string.toast_enable_silent, Toast.LENGTH_LONG).show()
-                            val intent = Intent(Settings.ACTION_NOTIFICATION_POLICY_ACCESS_SETTINGS)
-                            context.startActivityForResult(intent, 0)
-                            result.success(false)
-                            return
-                        }
-                    }
-
-                    // And we can enable the lesson notification mode manager.
-                    val intent = Intent(context, LessonNotificationModeManager::class.java)
-                    context.sendBroadcast(intent)
-                    result.success(true)
                 }
                 "activity.set_alarm" -> {
                     if (context !is Activity) {
@@ -151,9 +121,10 @@ class UnicaenTimetableApplication : FlutterApplication() {
     override fun onCreate() {
         super.onCreate()
         HeadlessTask.onInitialized { engine -> MethodChannel(engine.dartExecutor.binaryMessenger, CHANNEL).setMethodCallHandler { call, result -> handleMethod(call, result, applicationContext) } }
+    }
 
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            LessonNotificationModeManager.createChannel(this);
-        }
+    override fun attachBaseContext(base: Context) {
+        super.attachBaseContext(base)
+        MultiDex.install(this)
     }
 }
