@@ -1,51 +1,28 @@
-import 'dart:io';
-
-import 'package:admob_flutter/admob_flutter.dart';
 import 'package:background_fetch/background_fetch.dart';
-import 'package:catcher/catcher.dart';
 import 'package:ez_localization/ez_localization.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:hive/hive.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart' show MobileAds;
 import 'package:hive_flutter/hive_flutter.dart';
-import 'package:pedantic/pedantic.dart';
-import 'package:provider/provider.dart';
-import 'package:unicaen_timetable/credentials.dart';
+import 'package:rate_my_app/rate_my_app.dart';
 import 'package:unicaen_timetable/intro/scaffold.dart';
-import 'package:unicaen_timetable/model/home_cards.dart';
-import 'package:unicaen_timetable/model/lesson.dart';
+import 'package:unicaen_timetable/model/lessons/repository.dart';
+import 'package:unicaen_timetable/model/lessons/user/repository.dart';
+import 'package:unicaen_timetable/model/lessons/user/user.dart';
 import 'package:unicaen_timetable/model/settings/settings.dart';
-import 'package:unicaen_timetable/model/user.dart';
-import 'package:unicaen_timetable/pages/main_widget.dart';
+import 'package:unicaen_timetable/pages/page_container.dart';
 import 'package:unicaen_timetable/theme.dart';
-import 'package:unicaen_timetable/utils/widgets.dart';
+import 'package:unicaen_timetable/widgets/ensure_logged_in.dart';
 
 /// Hello world !
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  Admob.initialize();
+  MobileAds.instance.initialize();
   await Hive.initFlutter();
+  runApp(const ProviderScope(child: UnicaenTimetableRoot()));
 
-  if (kDebugMode) {
-    runApp(UnicaenTimetableApp());
-  } else {
-    CatcherOptions releaseConfig = CatcherOptions(SilentReportMode(), [
-      DiscordHandler(
-        Credentials.DISCORD_WEBHOOK,
-        enableDeviceParameters: false,
-        enableApplicationParameters: true,
-        enableCustomParameters: true,
-        enableStackTrace: true,
-        printLogs: true,
-      ),
-    ], customParameters: {
-      'platform': Platform.isAndroid ? 'Android' : 'iOS'
-    });
-    Catcher(rootWidget: UnicaenTimetableApp(), releaseConfig: releaseConfig);
-  }
-
-  unawaited(BackgroundFetch.registerHeadlessTask(headlessSyncTask));
+  BackgroundFetch.registerHeadlessTask(headlessSyncTask);
 }
 
 /// The headless synchronization task.
@@ -57,13 +34,13 @@ void headlessSyncTask(String taskId) async {
 
   User? user = await userRepository.getUser();
   if (user != null) {
-    LessonModel lessonModel = LessonModel();
+    LessonRepository lessonRepository = LessonRepository();
     SettingsModel settingsModel = SettingsModel();
 
     await settingsModel.initialize();
-    await lessonModel.initialize();
-    await lessonModel.synchronizeFromZimbra(
-      settingsModel: settingsModel,
+    await lessonRepository.initialize();
+    await lessonRepository.downloadLessons(
+      calendarUrl: settingsModel.calendarUrl,
       user: user,
     );
   }
@@ -73,89 +50,43 @@ void headlessSyncTask(String taskId) async {
 }
 
 /// The app first widget.
-class UnicaenTimetableApp extends StatefulWidget {
+class UnicaenTimetableRoot extends ConsumerStatefulWidget {
   /// The communication channel.
-  static const MethodChannel CHANNEL = MethodChannel('fr.skyost.timetable');
+  static const MethodChannel channel = MethodChannel('fr.skyost.timetable');
+
+  /// Creates a new Unicaen timetable app instance.
+  const UnicaenTimetableRoot({
+    Key? key,
+  }) : super(
+          key: key,
+        );
 
   @override
-  State<StatefulWidget> createState() => _UnicaenTimetableAppState();
+  ConsumerState createState() => _UnicaenTimetableRootState();
 }
 
 /// The app first widget state.
-class _UnicaenTimetableAppState extends State<UnicaenTimetableApp> {
-  /// The lesson model.
-  late LessonModel lessonModel;
-
-  /// The user repository.
-  late UserRepository userRepository;
-
-  /// The settings model.
-  late SettingsModel settingsModel;
-
+class _UnicaenTimetableRootState extends ConsumerState<UnicaenTimetableRoot> {
   @override
   void initState() {
     super.initState();
-
-    lessonModel = LessonModel();
-    userRepository = UserRepository();
-    settingsModel = SettingsModel();
-    _initialize();
+    initializeBackgroundTasks();
   }
 
   @override
   Widget build(BuildContext context) => EzLocalizationBuilder(
         delegate: const EzLocalizationDelegate(supportedLocales: [Locale('en'), Locale('fr')]),
-        builder: (context, ezLocalization) => MultiProvider(
-          providers: [
-            ChangeNotifierProvider<LessonModel>.value(value: lessonModel),
-            ChangeNotifierProvider<UserRepository>.value(value: userRepository),
-            ChangeNotifierProvider<SettingsModel>.value(value: settingsModel),
-            ChangeNotifierProvider<HomeCardsModel>(create: (_) => HomeCardsModel()..initialize(), lazy: false),
-          ],
-          child: Consumer<SettingsModel>(
-            builder: (context, settingsModel, child) => MaterialApp(
-              navigatorKey: Catcher.navigatorKey,
-              onGenerateTitle: (context) => EzLocalization.of(context)?.get('app_name') ?? 'Unicaen Timetable',
-              theme: UnicaenTimetableTheme.LIGHT.themeData,
-              darkTheme: UnicaenTimetableTheme.DARK.themeData,
-              themeMode: settingsModel.isInitialized ? settingsModel.themeEntry.value : ThemeMode.light,
-              routes: {
-                '/': (context) {
-                  if (!context.watch<LessonModel>().isInitialized || !context.watch<UserRepository>().isInitialized || !settingsModel.isInitialized) {
-                    return const Scaffold(
-                      body: CenteredCircularProgressIndicator(color: Colors.white),
-                      backgroundColor: Colors.indigo,
-                    );
-                  }
-
-                  return AppMainWidget();
-                },
-                '/intro': (_) => IntroScaffold(),
-              },
-              initialRoute: '/',
-              localizationsDelegates: ezLocalization.localizationDelegates,
-              supportedLocales: ezLocalization.supportedLocales,
-              localeResolutionCallback: ezLocalization.localeResolutionCallback,
-            ),
-          ),
-        ),
+        builder: (context, ezLocalization) => _UnicaenTimetableApp(ezLocalization: ezLocalization),
       );
 
   @override
   void dispose() {
-    lessonModel.dispose();
-    userRepository.dispose();
-    settingsModel.dispose();
     Hive.close();
     super.dispose();
   }
 
-  /// Initializes the models (and requests the tracking authorization).
-  Future<void> _initialize() async {
-    await lessonModel.initialize();
-    await userRepository.initialize();
-    await settingsModel.initialize();
-
+  /// Initializes the background tasks.
+  Future<void> initializeBackgroundTasks() async {
     await BackgroundFetch.configure(
       BackgroundFetchConfig(
         minimumFetchInterval: const Duration(days: 1).inMinutes,
@@ -165,11 +96,68 @@ class _UnicaenTimetableAppState extends State<UnicaenTimetableApp> {
         requiredNetworkType: NetworkType.ANY,
       ),
       (String taskId) async {
-        await lessonModel.synchronizeFromZimbra(settingsModel: settingsModel, user: await userRepository.getUser());
+        LessonRepository lessonRepository = ref.read(lessonRepositoryProvider);
+        await lessonRepository.initialize();
+
+        UserRepository userRepository = ref.read(userRepositoryProvider);
+        await userRepository.initialize();
+
+        SettingsModel settingsModel = ref.read(settingsModelProvider);
+        await settingsModel.initialize();
+
+        await lessonRepository.downloadLessons(calendarUrl: settingsModel.calendarUrl, user: await userRepository.getUser());
         BackgroundFetch.finish(taskId);
       },
     );
+  }
+}
 
-    await Admob.requestTrackingAuthorization();
+/// The app material widget.
+class _UnicaenTimetableApp extends ConsumerWidget {
+  /// The EzLocalization instance.
+  final EzLocalizationDelegate ezLocalization;
+
+  /// Creates a new Unicaen timetable app.
+  const _UnicaenTimetableApp({
+    required this.ezLocalization,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    SettingsModel settingsModel = ref.watch(settingsModelProvider);
+    return MaterialApp(
+      onGenerateTitle: (context) => EzLocalization.of(context)?.get('app_name') ?? 'Unicaen Timetable',
+      theme: UnicaenTimetableTheme.light.themeData,
+      darkTheme: UnicaenTimetableTheme.dark.themeData,
+      themeMode: settingsModel.isInitialized ? settingsModel.themeEntry.value : ThemeMode.light,
+      routes: {
+        '/': (_) => EnsureLoggedInWidget(
+          child: RateMyAppBuilder(
+            onInitialized: (context, rateMyApp) {
+              if (rateMyApp.shouldOpenDialog) {
+                rateMyApp.showRateDialog(
+                  context,
+                  title: context.getString('dialogs.rate.title'),
+                  message: context.getString('dialogs.rate.message'),
+                  rateButton: context.getString('dialogs.rate.button_rate').toUpperCase(),
+                  laterButton: context.getString('dialogs.rate.button_later').toUpperCase(),
+                  noButton: context.getString('dialogs.rate.button_no').toUpperCase(),
+                  ignoreNativeDialog: false,
+                );
+              }
+            },
+            builder: (context) => AnnotatedRegion<SystemUiOverlayStyle>(
+              value: SystemUiOverlayStyle.light.copyWith(systemNavigationBarColor: ref.watch(settingsModelProvider).resolveTheme(context).actionBarColor),
+              child: const PageContainer(),
+            ),
+          ),
+        ),
+        '/intro': (_) => const IntroScaffold(),
+      },
+      initialRoute: '/',
+      localizationsDelegates: ezLocalization.localizationDelegates,
+      supportedLocales: ezLocalization.supportedLocales,
+      localeResolutionCallback: ezLocalization.localeResolutionCallback,
+    );
   }
 }
