@@ -5,27 +5,31 @@ import 'package:unicaen_timetable/model/settings/entries/server/interval.dart';
 
 /// The server settings category.
 class ServerSettingsCategory extends SettingsCategory {
+  /// This category key.
+  static const String categoryKey = 'server';
+
   /// Creates a new server settings category instance.
   ServerSettingsCategory()
       : super(
-          key: 'server',
+          key: categoryKey,
           icon: Icons.wifi,
-        ) {
-    addEntry(IntervalSettingsEntry(keyPrefix: key));
-    addEntry(SettingsEntry<String>(
-      keyPrefix: key,
-      key: 'server',
-      value: 'https://webmail.unicaen.fr',
-    ));
-    addEntry(SettingsEntry<String>(
-      keyPrefix: key,
-      key: 'calendar',
-      value: 'Emploi du temps',
-    ));
-    addEntry(SettingsEntry<String>(
-      keyPrefix: key,
-      key: 'additional_parameters',
-      value: '',
-    ));
-  }
+          entries: [
+            IntervalSettingsEntry(keyPrefix: categoryKey),
+            SettingsEntry<String>(
+              categoryKey: categoryKey,
+              key: 'server',
+              value: 'https://webmail.unicaen.fr',
+            ),
+            SettingsEntry<String>(
+              categoryKey: categoryKey,
+              key: 'calendar',
+              value: 'Emploi du temps',
+            ),
+            SettingsEntry<String>(
+              categoryKey: categoryKey,
+              key: 'additional_parameters',
+              value: '',
+            ),
+          ],
+        );
 }
